+++ packages/side-effects/side-effect-title/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 1.0.0-beta.7
 
 * Add a `delimiter` option (defaults to an empty string `''`).
 * Removed source maps from `dist`